requirements.txt
asgiref==3.3.4
boto3==1.16.43
botocore==1.19.43
certifi==2020.12.5
chardet==3.0.4
dataclasses==0.6
decorator==4.4.2
dj-database-url==0.5.0
Django==3.2.3
django-environ==0.4.5
django-heroku==0.3.1
future==0.18.2
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
numpy==1.19.4
pandas==1.1.5
patsy==0.5.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
PyHive==0.5.2
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
retry==0.9.2
s3transfer==0.3.3
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
statsmodels==0.12.1
typing-extensions==3.10.0.0
urllib3==1.25.11
wh==1.2.0
whitenoise==5.2.0
environ~=1.0
Pillow
django_tables2
django_filter